Update practise for markdown
Practiced the use of link markdown to display image from the cloud with the link of the image.But the image was not displayed because the size was too big. Then used regular html to display image and it was displaying because with use of html, the image was compressed.